pyproject.toml
[tool.poetry]
name = "convert-to-pdf"
version = "0.1.0"
description = "project to convert files to pdf"
authors = ["KLYMENKORUS <klimenkoruslan11@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.1"
uvicorn = "^0.23.1"
pydantic = {extras = ["email"], version = "^2.1.1"}
aerich = "^0.7.2"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.3"}
asyncpg = "^0.28.0"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.24.1"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
celery = {extras = ["redis"], version = "^5.3.1"}
flower = "^2.0.1"



[tool.aerich]
tortoise_orm = "app.database.config.settings.DB_CONFIG"
location = "app/database/migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
